# Add metrics to ASP.NET Core