--- a/src/Hosting/Abstractions/src/Microsoft.AspNetCore.Hosting.Abstractions.csproj
+++ b/src/Hosting/Abstractions/src/Microsoft.AspNetCore.Hosting.Abstractions.csproj
@@ -16,7 +16,7 @@
   </ItemGroup>
 
   <!-- Temporary hack to make prototype Metrics DI integration types available -->
-  <!-- TODO: Remove when Metrics DI intergration package is available -->
+  <!-- TODO: Remove when Metrics DI intergration package is available https://github.com/dotnet/aspnetcore/issues/47618 -->
   <ItemGroup>
     <InternalsVisibleTo Include="Microsoft.AspNetCore.Hosting" />
     <InternalsVisibleTo Include="Microsoft.AspNetCore.Hosting.Tests" />