--- a/src/Hosting/Hosting/src/GenericHost/SlimWebHostBuilder.cs
+++ b/src/Hosting/Hosting/src/GenericHost/SlimWebHostBuilder.cs
@@ -53,6 +53,9 @@ internal sealed class SlimWebHostBuilder : WebHostBuilderBase, ISupportsStartup
             services.TryAddSingleton<IHttpContextFactory, DefaultHttpContextFactory>();
             services.TryAddScoped<IMiddlewareFactory, MiddlewareFactory>();
             services.TryAddSingleton<IApplicationBuilderFactory, ApplicationBuilderFactory>();
+
+            services.AddMetrics();
+            services.TryAddSingleton<HostingMetrics>();
         });
     }