--- a/src/Hosting/Hosting/src/Internal/HttpMetricsTagsFeature.cs
+++ b/src/Hosting/Hosting/src/Internal/HttpMetricsTagsFeature.cs
@@ -1,0 +1,13 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using Microsoft.AspNetCore.Http.Features;
+
+namespace Microsoft.AspNetCore.Hosting;
+
+internal sealed class HttpMetricsTagsFeature : IHttpMetricsTagsFeature
+{
+    public ICollection<KeyValuePair<string, object?>> Tags => TagsList;
+
+    public List<KeyValuePair<string, object?>> TagsList { get; } = new List<KeyValuePair<string, object?>>();
+}