… src/Hosting/Hosting/src/Internal/WebHost.cs
… src/Hosting/Hosting/test/HostingApplicationDiagnosticsTests.cs
… src/Hosting/Hosting/test/HostingApplicationTests.cs
… src/Hosting/Hosting/test/Internal/HostingEventSourceTests.cs
… src/Http/Http.Abstractions/src/Metadata/IRouteDiagnosticsMetadata.cs
… src/Http/Http.Abstractions/src/PublicAPI.Unshipped.txt
… src/Http/Http.Features/src/IHttpMetricsTagsFeature.cs
… src/Http/Http.Features/src/PublicAPI.Unshipped.txt
… src/Http/Routing/src/Patterns/RoutePattern.cs
… src/Http/Routing/src/RouteEndpointBuilder.cs
… src/Http/Routing/test/UnitTests/Builder/RequestDelegateEndpointRouteBuilderExtensionsTest.cs
… src/Http/Routing/test/UnitTests/CompositeEndpointDataSourceTest.cs
… src/Http/Routing/test/UnitTests/RouteEndpointBuilderTest.cs
… src/Http/samples/MinimalSample/Program.cs
… src/Middleware/Diagnostics/src/DeveloperExceptionPage/DeveloperExceptionPageMiddlewareImpl.cs
… src/Middleware/Diagnostics/src/DiagnosticsTelemetry.cs
… src/Middleware/Diagnostics/src/ExceptionHandler/ExceptionHandlerMiddlewareImpl.cs
… src/Middleware/Diagnostics/test/UnitTests/DeveloperExceptionPageMiddlewareTest.cs
… src/Middleware/Diagnostics/test/UnitTests/ExceptionHandlerTest.cs
… src/Middleware/Middleware.slnf
… src/Middleware/RateLimiting/samples/RateLimitingSample/RateLimitingSample.csproj
… src/Mvc/Mvc.Core/test/ApplicationModels/EndpointMetadataProviderTest.cs
… src/Mvc/Mvc.Razor.RuntimeCompilation/src/RuntimeViewCompiler.cs
… src/Servers/Connections.Abstractions/src/Features/IConnectionMetricsTagsFeature.cs
… src/Servers/Connections.Abstractions/src/PublicAPI/net462/PublicAPI.Unshipped.txt
… src/Servers/Connections.Abstractions/src/PublicAPI/net8.0/PublicAPI.Unshipped.txt
… src/Servers/Connections.Abstractions/src/PublicAPI/netstandard2.0/PublicAPI.Unshipped.txt
… src/Servers/Connections.Abstractions/src/PublicAPI/netstandard2.1/PublicAPI.Unshipped.txt
… src/Servers/Kestrel/Core/src/Internal/ConnectionDispatcher.cs
… src/Servers/Kestrel/Core/src/Internal/Http/Http1Connection.cs