This PR contains several changes:

* Prototype for System.Diagnostics.Metrics integration with DI (plan is for a new Microsoft.Extensions.Metrics package)
  * This is providing value by making metrics testable but isn't publically available. The **temporary** workaround is to put types in ASP.NET Core, make them internal, then use `InternalsVisibleTo`. An alternative idea could be to put these types in a new assembly and keep it internal by not publishing ref pack, but that's more work, and this is just a temporary message.
  * The DI API surface might change, but I expect reacting to the public API won't be a big task.
* Adding metrics to ASP.NET Core. This means adding metrics to places that have event counters today. There is an issue that specifies all the counters added in ASP.NET Core here: https://github.com/dotnet/aspnetcore/issues/47536
  * Add metrics to HTTP hosting. Two new public APIs are used here:
    * `IHttpMetricsTagsFeature`. API issue: https://github.com/dotnet/aspnetcore/issues/47493
    * `IRouteDiagnosticsMetadata`. API issue: https://github.com/dotnet/aspnetcore/issues/47492
  * Add metrics to Kestrel. One new public API is used here:
    * `IConnectionMetricsTagsFeature`. API issue: https://github.com/dotnet/aspnetcore/issues/47493
  * Add metrics to SignalR.