@mitchdenny 2023-03-23
What are we going to do around naming conventions for meters/counters? It looks like we'll adopt the names of event sources where they exist, but eventually we'll be breaking new ground and want to add some metrics where there wasn't previous instrumentation. What is the naming convention that we should be adopting?