@davidfowl 2023-03-30
Make sure you benchmark this