@danmoseley 2023-03-30
Maybe paste your https://gist.github.com/JamesNK/d07cf4619f01e17d392afc9520e6bc18 in here for posterity?