@JamesNK 2023-04-06
> @JamesNK have you tried to visualize these counters in any tools like app insights or grafana?

Not yet.