@JamesNK 2023-04-06
> Consider starting from Grafana, and once the dashboard looks good, export the JSON file (maybe contribute the JSON file to Grafana dashboard registry [grafana.com/grafana/dashboards](https://grafana.com/grafana/dashboards/)).

That's the plan. @davidfowl and I discussed making a semi-official Grafana dashboard for asp.net core.