@JamesNK 2023-04-12
Ran microbenchmarks in Kestrel. No noticeable change.

HTTP/1.1

```
Main

|               Method |     Mean |   Error |  StdDev |        Op/s | Ratio | RatioSD | Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------------------- |---------:|--------:|--------:|------------:|------:|--------:|------:|------:|------:|----------:|
| PlaintextTechEmpower | 216.2 ns | 2.18 ns | 2.04 ns | 4,626,120.1 |  1.00 |    0.00 |     - |     - |     - |         - |
|           LiveAspNet | 407.4 ns | 2.16 ns | 1.81 ns | 2,454,805.0 |  1.88 |    0.02 |     - |     - |     - |      56 B |

PR

|               Method |     Mean |   Error |  StdDev |        Op/s | Ratio | RatioSD | Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------------------- |---------:|--------:|--------:|------------:|------:|--------:|------:|------:|------:|----------:|
| PlaintextTechEmpower | 217.9 ns | 1.88 ns | 1.76 ns | 4,590,132.1 |  1.00 |    0.00 |     - |     - |     - |         - |
|           LiveAspNet | 410.2 ns | 1.38 ns | 1.22 ns | 2,437,618.5 |  1.88 |    0.02 |     - |     - |     - |      56 B |
```

HTTP/2

```
main

|      Method | ResponseDataLength | NumCookies |     Mean |     Error |    StdDev |      Op/s | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------ |------------------- |----------- |---------:|----------:|----------:|----------:|------:|------:|------:|----------:|
| MakeRequest |                  0 |          0 | 4.417 us | 0.0610 us | 0.0541 us | 226,383.7 |     - |     - |     - |     376 B |
| MakeRequest |                  0 |          1 | 4.511 us | 0.0667 us | 0.0624 us | 221,672.7 |     - |     - |     - |     376 B |
| MakeRequest |                  0 |          3 | 4.994 us | 0.0486 us | 0.0380 us | 200,233.7 |     - |     - |     - |     608 B |

PR

|      Method | ResponseDataLength | NumCookies |     Mean |     Error |    StdDev |      Op/s | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------ |------------------- |----------- |---------:|----------:|----------:|----------:|------:|------:|------:|----------:|
| MakeRequest |                  0 |          0 | 4.356 us | 0.0723 us | 0.0640 us | 229,578.6 |     - |     - |     - |     376 B |
| MakeRequest |                  0 |          1 | 4.515 us | 0.0642 us | 0.0536 us | 221,462.7 |     - |     - |     - |     376 B |
| MakeRequest |                  0 |          3 | 4.709 us | 0.0428 us | 0.0357 us | 212,347.8 |     - |     - |     - |     608 B |
```