@JamesNK 2023-04-13
> Did you run any tech empower benchmarks?

PlaintextTechEmpower microbenchmark.